Created `FileReader` actor to handle all the file opening, splitting and flow control logic. Should make it available as an initial build command to the pipeline next.
This actor holds a bunch of mutable variables. The number of file chunks that will be produced is set only by the `SplitFile` message, and not when the actor is created. So is the filename.

When the message is received, the actor creates a file chunk iterator that is stored in `chunkIterator`. This iterator is only accessed by the `sendChunks` method (yeas, they should all be private, I dunno).

Other two variables keep track of how many chunks have been sent to processing, and how many have finished.