A few major changes due to the `Reducer` router now being based on `BalancingPool` instead of `ConsistentHashing`.
The change gives a 4x speedup on the benchmark. We now have a termination based control logic too, so the `Decimator` is about to retire. And finally, we are now accumulating counters in the supervisor, since keys can be duplicated. Gotta figure out now if the mappers could be writing directly to the router, so we never have a single actor as a bottleneck.